requirements.txt
click==8.1.3
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
ptyprocess==0.7.0
ruuvitag-sensor==1.2.0
Rx==1.6.1
Werkzeug==2.1.2
python-dotenv==0.20.0